+++ firmware/lib/SoftwareSerialPort/softwareserialport.cpp
@@ -3,8 +3,8 @@
 SoftwareSerialPort::SoftwareSerialPort(SoftwareSerial *serialPort,
                                        uint8_t rxPin, 
                                        uint8_t txPin, 
-                                       long long baudRate, 
-                                       long long timeout,
+                                       uint32_t baudRate, 
+                                       uint32_t timeout,
                                        bool enabled,
                                        const char *lineEnding) :
     ByteStream(serialPort, rxPin, txPin, baudRate, timeout, enabled, lineEnding),
